move d2common/cache.go into sub-package d2common/d2cache
